% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj.R
\docType{data}
\name{adj}
\alias{adj}
\title{An adjacent information for example dataset.}
\format{
A data frame with 40 rows and 2 variables:
\describe{
\item{areaNO}{area number}
\item{adjNO}{adjacent area number}
}
}
\source{
none
}
\usage{
adj
}
\description{
A dataset is the adjacent information for 10 areas.
}
\keyword{datasets}
