% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGFL_cda1.R
\name{GGFL.cda1}
\alias{GGFL.cda1}
\title{Sub-function for GGFL}
\usage{
GGFL.cda1(M, c, Lambda, B, Beta0, thres = 1e-05)
}
\arguments{
\item{M}{k × k positive definite matrix}

\item{c}{k-dimensional vector}

\item{Lambda}{r-dimensional vector which all elements are positive}

\item{B}{r × k matrix}

\item{Beta0}{k-dimensional initial vector}

\item{thres}{threshold for convergence judgement}
}
\value{
solution: vector of the minimizer

type: "join" or "disjoin"

idx: index of parameter when type=join
}
\description{
\code{GGFL.cda1} This is required to execute "GGFL.cda" or "GGFLa.cda"
}
\examples{
#GGFL.cda1(M, c, Lambda, B, Beta0)
}
