% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGFL_cda.R
\name{GGFL}
\alias{GGFL}
\title{Coordinate optimization for GGFL}
\usage{
GGFL(
  y,
  X,
  area,
  adj,
  Lambda = "default",
  lambda.type = NULL,
  adaptive = TRUE,
  weight = NULL,
  standardize = TRUE,
  MPinv = FALSE,
  tol = 1e-05,
  progress = FALSE,
  out.all = FALSE
)
}
\arguments{
\item{y}{vector of an objective variable}

\item{X}{matrix of explanatory variables without intercept}

\item{area}{vector of area labels for each individual}

\item{adj}{dataframe of adjacent information of which
the labels for the first and second column are respectively "area" and "adj"}

\item{Lambda}{candidates of tuning parameter
if "default", the candidates are defined following the paper;
if "uniform", the candidates are defined by uniformly dividing;
if vector/scalar, the candidates are defined by following lambda.type}

\item{lambda.type}{option when Lambda is numeric;
if "value", Lambda is searching points;
if "rate", lam.max*Lambda is searching points}

\item{adaptive}{option only when weight=NULL; if true, adaptive penalty is adopted}

\item{weight}{list of penalty weight (default is TRUE)}

\item{standardize}{if TRUE, y and X are standardized in the sense of norm}

\item{MPinv}{if TRUE, the ordinary least square estimator is calculated by the Moore–Penrose inverse matrix}

\item{tol}{tolerance for convergence}

\item{progress}{If TRUE, progress is displayed}

\item{out.all}{if TRUE, results for all tuning parameters are output;
if FALSE, results for only the optimal tuning parameter are output}
}
\value{
idx: the index of the optimal tuning parameter

Coef: list of GGFL estimates

rss: vector or scalar of residual sum of squares

df: vector or scalar of degrees of freedom

msc: vector or scalar of values of model selection criterion

r2: vector or scalar of coefficient of determination

mer: vector or scalar of median error rate

lambda: vector or scalar of tuning parameters

weight: list of penalty weight

gr.labs: list of labels for groups

time: runtime (s)

coef.opt: matrix of GGFL estimates

coef.lse: matrix of OLS estimates

coef.max: matrix of estimates when all groups are equal

pred: vecter of predictive values for coef.opt

pred.lse: vecter of predictive values for coef.lse

pred.max: vecter of predictive values for coef.max
}
\description{
\code{GGFL} solving GGFL optimization problem via coordinate descent algorithm
}
\examples{
#GGFL(y, X, area, adj)
}
