% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGFLa_cda.R
\name{GGFLa.cda}
\alias{GGFLa.cda}
\title{Coordinate optimization for GGFL}
\usage{
GGFLa.cda(
  yli,
  Xli,
  Zli,
  D,
  thres = 1e-05,
  Lambda = "default",
  MPinv = FALSE,
  progress = FALSE,
  out.all = FALSE
)
}
\arguments{
\item{yli}{list of group-wise vectors of an objective variable}

\item{Xli}{list of group-wise matrixes of explanatory variables which are took GGFL penalty}

\item{Zli}{list of group-wise matrixes of explanatory variables which are not took GGFL penalty}

\item{D}{list of adjacency relations}

\item{thres}{threshold for convergence judgement}

\item{Lambda}{candidates of tuning parameter
if "default", the candidates are defined following the paper;
if "uniform", the candidates are defined by uniformly dividing}

\item{MPinv}{if TRUE, the ordinary least square estimator is calculated by the Moore–Penrose inverse matrix}

\item{progress}{If TRUE, progress is displayed}

\item{out.all}{if TRUE, results for all tuning parameters are output;
if FALSE, results for only the optimal tuning parameter are output}
}
\value{
coef: list of GGFL estimates

rss: vector or scalar of residual sum of squares

df: vector or scalar of degrees of freedom

msc: vector or scalar of values of model selection criterion

r2: vector or scalar of coefficient of determination

mer: vector or scalar of median error rate

lambda: vector or scalar of tuning parameters

weight: list of penalty weight

gr.labs: list of labels for groups

time: runtime (s)

coef.opt: matrix of GGFL estimates

coef.lse: matrix of OLS estimates

coef.max: matrix of estimates when all groups are equal

pred: vecter of predictive values for coef.opt

pred.lse: vecter of predictive values for coef.lse

pred.max: vecter of predictive values for coef.max
}
\description{
\code{GGFLa.cda} More general version of Coordinate optimization for GGFL.
This allows to use explanatory variables which are not took GGFL penalty.
}
\examples{
#GGFLa.cda(yli, Xli, Zli, D)
}
