% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGFLa_cda.R
\name{pGGFL}
\alias{pGGFL}
\title{Coordinate optimization for GGFL}
\usage{
pGGFL(
  y,
  X,
  Z,
  group,
  adj,
  tol = 1e-05,
  Lambda = "default",
  alpha = NULL,
  standardize = TRUE,
  MPinv = FALSE,
  maxit = 500,
  progress = FALSE,
  out.all = FALSE,
  cwu.control = NULL
)
}
\arguments{
\item{y}{vector of an objective variable}

\item{X}{matrix of explanatory variables without intercept for local estimation
(for GGFL penalty)}

\item{Z}{matrix of explanatory variables without intercept for global estimation}

\item{group}{numeric vector of group labels for each individual}

\item{adj}{data.frame or matrix with two columns of adjacent information}

\item{tol}{tolerance for convergence}

\item{Lambda}{candidates of tuning parameter
if \code{"default"}, the candidates are defined following the paper;
if \code{"uniform"}, the candidates are defined by uniformly dividing}

\item{alpha}{a value expressing penalty strength for EGCV criterion;
default is \code{NULL} which corresponds to \code{log(n)}}

\item{standardize}{if \code{TRUE}, \code{X} is standardized in the sense of norm}

\item{MPinv}{if TRUE, the ordinary least squares estimator is calculated by the Moore-Penrose inverse matrix}

\item{maxit}{iteration limit}

\item{progress}{If \code{TRUE}, progress is displayed of the form
"lambda index_iteration number_convergence criterion_objective function"}

\item{out.all}{if \code{TRUE}, results for all tuning parameters are output;
if \code{FALSE}, results for only the optimal tuning parameter are output}

\item{cwu.control}{a list object of parameters for controlling the coordinate wise update
which can has the two elements \code{conv.check} and \code{tol};
\code{conv.check} is one of "coef" (default) and "grad" and solution convergence is
determined based on the coefficients or gradient, respectively;
\code{tol} is a tolerance for convergence of which the default is 1e-5}
}
\value{
a list object which has the following elements:
\item{coefficients.local}{a list object for \code{X} which has
\code{GGFL}: matrix of GGFL estimates; \code{OLS}: matrix of OLS estimates;
\code{MAX} matrix of estimates when all groups are equal
}

\item{coefficients.global}{a list object for \code{Z} which has
\code{GGFL}: matrix of GGFL estimates; \code{OLS}: matrix of OLS estimates;
\code{MAX} matrix of estimates when all groups are equal
}

\item{fitted.values}{a matrix of fitted.values corresponding to
each element of \code{coefficients}
}

\item{summary}{a data.frame with results of the optimal solution}

\item{weight}{a list of penalty weight}

\item{cluster}{a list of cluster}

\item{cluster.labels}{a vector expressing which cluster each group is in}

\item{cwu.control}{\code{cwu.control}}

\item{all.coef}{a list with all candidates of GGFL estimates}

\item{all.cluster}{a matrix of \code{cluster}s for all candidates}

\item{log}{results for all candidates}
}
\description{
\code{pGGFL} More general version of Coordinate optimization for GGFL.
Only partial explanatory variables are penalized. (v1.0.1)
}
\examples{
#pGGFL(y, X, Z, group, adj)
}
