% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGFL_cda.R
\name{GGFL}
\alias{GGFL}
\title{Coordinate optimization for GGFL}
\usage{
GGFL(
  y,
  X,
  area,
  adj,
  Lambda = "default",
  lambda.type = NULL,
  adaptive = TRUE,
  weight = NULL,
  alpha = NULL,
  standardize = TRUE,
  MPinv = FALSE,
  tol = 1e-05,
  maxit = 500,
  progress = FALSE,
  out.all = FALSE
)
}
\arguments{
\item{y}{vector of an objective variable}

\item{X}{matrix of explanatory variables without intercept}

\item{area}{vector of area labels for each individual}

\item{adj}{dataframe of adjacent information of which
the labels for the first and second column are respectively "area" and "adj"}

\item{Lambda}{candidates of tuning parameter
if \code{"default"}, the candidates are defined following the paper;
if \code{"uniform"}, the candidates are defined by uniformly dividing;
if vector/scalar, the candidates are defined by following \code{lambda.type}}

\item{lambda.type}{option when \code{Lambda} is numeric;
if "value", \code{Lambda} is searching points;
if "rate", \code{lam.max*Lambda} is searching points}

\item{adaptive}{option only when \code{weight=NULL}; if \code{TRUE}, adaptive penalty is adopted}

\item{weight}{list of penalty weight (default is \code{NULL})}

\item{alpha}{a value expressing penalty strength for EGCV criterion;
default is \code{NULL} which corresponds to \code{log(n)}}

\item{standardize}{if \code{TRUE}, \code{y} and \code{X} are standardized in the sense of norm}

\item{MPinv}{if \code{TRUE}, the ordinary least squares estimator is calculated by the Moore-Penrose inverse matrix}

\item{tol}{tolerance for convergence}

\item{maxit}{iteration limit}

\item{progress}{If \code{TRUE}, progress is displayed}

\item{out.all}{if \code{TRUE}, results for all tuning parameters are output;
if \code{FALSE}, results for only the optimal tuning parameter are output}
}
\value{
a list object which has the following elements:
\item{idx}{the index of the optimal tuning parameter}

\item{Coef}{list of GGFL estimates}

\item{rss}{vector or scalar of residual sum of squares}

\item{df}{vector or scalar of degrees of freedom}

\item{msc}{vector or scalar of values of EGCV criterion}

\item{r2}{vector or scalar of coefficient of determination}

\item{mer}{vector or scalar of median error rate}

\item{lambda}{vector or scalar of tuning parameters}

\item{weight}{list of penalty weight}

\item{gr.labs}{list of labels for groups}

\item{time}{runtime (s)}

\item{coefGGFL}{matrix of GGFL estimates}

\item{coefOLS}{matrix of OLS estimates}

\item{coefMAX}{matrix of estimates when all groups are equal}

\item{fitGGFL}{vecter of fitted values for \code{coefGGFL}}

\item{fitOLS}{vecter of fitted values for \code{coefOLS}}

\item{fitMAX}{vecter of fitted values for \code{coefMAX}}
}
\description{
\code{GGFL} solving GGFL optimization problem via coordinate descent algorithm
}
\examples{
#GGFL(y, X, area, adj)
}
