% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGFLa_cda.R
\name{pGGFL}
\alias{pGGFL}
\title{Coordinate optimization for GGFL}
\usage{
pGGFL(
  yli,
  Xli,
  Zli,
  D,
  tol = 1e-05,
  Lambda = "default",
  alpha = NULL,
  MPinv = FALSE,
  maxit = 500,
  progress = FALSE,
  out.all = FALSE,
  cwu.control = NULL
)
}
\arguments{
\item{yli}{list of group-wise vectors of an objective variable}

\item{Xli}{list of group-wise matrixes of explanatory variables which are took GGFL penalty}

\item{Zli}{list of group-wise matrixes of explanatory variables which are not took GGFL penalty}

\item{D}{list of adjacency relations}

\item{tol}{tolerance for convergence}

\item{Lambda}{candidates of tuning parameter
if \code{"default"}, the candidates are defined following the paper;
if \code{"uniform"}, the candidates are defined by uniformly dividing}

\item{alpha}{a value expressing penalty strength for EGCV criterion;
default is \code{NULL} which corresponds to \code{log(n)}}

\item{MPinv}{if TRUE, the ordinary least squares estimator is calculated by the Moore-Penrose inverse matrix}

\item{maxit}{iteration limit}

\item{progress}{If \code{TRUE}, progress is displayed}

\item{out.all}{if \code{TRUE}, results for all tuning parameters are output;
if \code{FALSE}, results for only the optimal tuning parameter are output}

\item{cwu.control}{a list object of parameters for controlling the coordinate wise update
which can has the two elements \code{conv.check} and \code{tol};
\code{conv.check} is one of "coef" (default) and "grad" and solution convergence is
determined based on the coefficients or gradient, respectively;
\code{tol} is a tolerance for convergence of which the default is 1e-5}
}
\value{
a list object which has the following elements:
\item{coef1}{list of GGFL estimates for penalized variables}

\item{coef2}{list of GGFL estimates for non-penalized variables}

\item{rss}{vector or scalar of residual sum of squares}

\item{df}{vector or scalar of degrees of freedom}

\item{msc}{vector or scalar of values of model selection criterion}

\item{r2}{vector or scalar of coefficient of determination}

\item{mer}{vector or scalar of median error rate}

\item{lambda}{vector or scalar of tuning parameters}

\item{weight}{list of penalty weight}

\item{gr.labs}{list of labels for groups}

\item{time}{runtime (s)}

\item{coef1GGFL}{matrix of GGFL estimates for penalized variables}

\item{coef1OLS}{matrix of OLS estimates for penalized variables}

\item{coef1MAX}{matrix of estimates for penalized variables when all groups are equal}

\item{coef2GGFL}{matrix of GGFL estimates for non-penalized variables}

\item{coef2OLS}{matrix of OLS estimates for non-penalized variables}

\item{coef2MAX}{matrix of estimates for non-penalized variables when all groups are equal}

\item{fitGGFL}{vector of fitted values for \code{coef1GGFL} and \code{coef2GGFL}}

\item{fitOLS}{vector of fitted values for \code{coef1OLS} and \code{coef2OLS}}

\item{fitMAX}{vector of fitted values for \code{coef1MAX} and \code{coef2MAX}}

\item{cwu.control}{\code{cwu.control}}
}
\description{
\code{pGGFL} More general version of Coordinate optimization for GGFL.
Only partial explanatory variables are penalized. (v0.2.0)
}
\examples{
#pGGFL(yli, Xli, Zli, D)
}
